Add basic update user support
Support simple updating of the user, however prevent changing of the
email as it requires either a multi item transaction or a secondary
index with a different attribute used as a primary key (e.g. Id).

Could also be done on a small scan using scan/list with a filter
expression, or via multiple tables.