Reintroduce Id and implement data store support
Switch back to using a separate Id identifier for users that can be
stored internally and permits updating of the Email attribute.

Provide separate calls for GetById and GetByEmail that relies on a
second item created for each user that ensures the email address is
unique to each user by using it as a primary id.

Update all impacted layers to switch to the correct new Id, and update
the list behaviour to correctly perform filtering now that multiple
objectTypes exist by default.